// src/app/api/generated/model/updateCardDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Boards API
 * REST API for Boards, Spring Boot version
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateCardDto {
  /** @minLength 1 */
  title: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  description?: string;
}